feat: Add GeneLinkChips Component for Outlinks in Curation Modal
Implemented the GeneLinkChips component to provide outlinks to ClinGen, GenCC, and OMIM in the curation modal. The component is designed to be reusable and customizable, allowing for the display of different links based on input parameters such as the HGNC id, gene symbol, and OMIM number. Links are presented as clickable chips, enhancing the user interface. Enhanced the component with inline documentation and JSDoc-style comments for clarity and maintainability. Integrated the component into the CurationModal's header and provided styling for a more intuitive user interaction. This feature addresses and closes Feature request: Add Outlinks to ClinGen and OMIM in curation modal #83.

closes  #83